Move defense position stuff to evaluation
