--- src/software/ai/evaluation/defense_lane.cpp
@@ -1,1 +1,0 @@
-#include "software/ai/evaluation/defense_lane.h"
