--- src/software/ai/evaluation/defense_lane.h
@@ -1,20 +1,0 @@
-#pragma once
-
-#include "software/geom/point.h"
-#include "software/geom/segment.h"
-
-struct DefenseLane
-{
-    Segment lane;
-
-    unsigned int expected_threat;
-};
-
-struct DefensePosition
-{
-    Point position;
-
-    unsigned int expected_threat;
-
-    bool is_crease_defense;
-};
